+++ package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "bootstrap": "^4.3.1",
     "dotenv": "^7.0.0",
+    "moment": "^2.24.0",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
     "react-scripts": "2.1.8",
